refactor(agent): restore live terminal with local composer
