+++ CHANGELOG.md
@@ -9,30 +9,14 @@ Entries reference the issue that motivated them.
 
 ### Added
 
-- Opening an Agent now lands on Monitor: a color-preserving snapshot of its
-  latest screen with an honest capture time. Attach remains available as a
-  full-screen toolbar action for realtime interaction; leaving it closes the
-  live terminal and refreshes Monitor once. Agent Notification taps land on
-  Monitor as well. (#179)
-
-- Monitor now follows a Working Agent on a two-second foreground cadence and
-  stops polling as soon as the Agent becomes Idle, Done, or Blocked. Status
-  pushes and pull-to-refresh still update the snapshot; pinned output follows
-  automatically, while scrolling up holds position behind a New Output jump.
-  (#180)
-
-- Monitor keeps a local scrollback. Scrolling up through what the Agent did
-  is served from the on-device cache; reaching the top while the Agent is
-  idle backfills up to herdr's 1,000-line capture window, the one loading
-  indicator in the surface. While the Agent works, the top of the cache says
-  history is unavailable instead of spinning. Reads that cannot be
-  reconciled with the cache leave an explicit gap marker, and the oldest
-  capturable line is marked as the beginning of the captured history. (#181)
-- Monitor now includes a local Composer. Drafting makes no network requests;
+- Agent detail now combines the complete live libghostty terminal with a local
+  Composer. The terminal preserves TUI rendering, scrollback, and PTY resize,
+  but no longer accepts direct keyboard, paste, or mouse-reporting input.
+  Drafting makes no network requests;
   Send delivers the complete message once and shows delivery, Agent work, and
   Done states from acknowledgments and status pushes. Failed messages can be
   retried or returned to the draft without losing text, and drafts survive
-  Attach, backgrounding, and Host reconnects. The existing Agent switcher now
+  terminal recovery, backgrounding, and Host reconnects. The Agent switcher
   sits at the bottom of the Composer, keeping session changes and the keyboard
   toggle within reach without reopening the Console. (#182)
 
@@ -41,17 +25,6 @@ Entries reference the issue that motivated them.
   and its secondary sources, OpenSSL, Ghostty and its stack, the monospaced
   fonts, and the rest of the audited inventory. The catalogue fails loudly
   when a notice is missing or malformed. (#161; PR #170)
-
-- The Keys keyboard has a Skills pane: the skills and custom slash commands
-  installed on the Host for the agent you're attached to — the project's own
-  and the global ones — listed by name and description, one tap inserting
-  the invocation into the terminal without sending it, a long press
-  previewing the full description. Covered kinds and
-  their mechanisms: Claude Code Agent Skills (`/name`), Codex Agent Skills
-  (`$name`), OpenCode commands (`/name`), and Pi skills (`/skill:name`) plus
-  prompt templates (`/name`). The list is probed once per connection and
-  refreshable from the pane; agents of a kind without a known mechanism
-  don't show the tab.
 
 ### Changed
 
@@ -86,18 +59,6 @@ Entries reference the issue that motivated them.
   `authorized_keys` keeps working.
 
 ### Fixed
-
-- Monitor snapshots no longer scroll horizontally. Their text now stays
-  constrained to the available screen width while captured history remains
-  vertically scrollable. (#179)
-
-- Monitor now presents a stable status summary and high-contrast snapshot as an
-  Agent response inside one conversation-style flow with locally sent messages,
-  a compact floating Composer, and a toolbar path into the live terminal.
-  Agent status now sits directly above the Composer, and the trailing input
-  region from the Agent TUI no longer duplicates the native message field.
-  Status cards, terminal shortcut controls, and oversized empty-state icons no
-  longer crowd the conversation. (#179; #182)
 
 - Malformed herdr API error responses that carry an empty id now fail the
   originating request immediately with the server's error instead of hanging
